ice merge with defaultsegment and UI update
