+++ components/Game/Segment.tsx
@@ -5,7 +5,6 @@ import itemStore from "@/components/Game/store";
 import { useRouter } from "next/router";
 import styles from "./styles.module.css";
 import DefaultSegment from "./SegmentState/defaultSegment";
-import Ice from "./SegmentState/ice";
 import Rocket from "./SegmentState/rocket";
 import Magnet from "./SegmentState/magnet";
 import { useRecoilState, useRecoilValue } from "recoil";
@@ -24,7 +23,7 @@ function Segment({ i, auth, videoId, peerxy, dataChannel, segmentState, isRightC
   const peerWait = useRecoilValue(peerWaitState);
   return (
     <>
-      {(auth ? myWait === 0 : peerWait === 0) && (segmentState === "default" || segmentState === "lip" || segmentState === "twirl") && (
+      {(auth ? myWait === 0 : peerWait === 0) && (segmentState === "default" || segmentState === "lip" || segmentState === "twirl" || segmentState === "ice") && (
         <DefaultSegment
           key={`default_${i}`}
           i={i}
@@ -38,9 +37,6 @@ function Segment({ i, auth, videoId, peerxy, dataChannel, segmentState, isRightC
       )}
       {(auth ? myWait === 9 : peerWait === 9) && segmentState === "rocket" && (
         <Rocket key={`rocket_${i}`} i={i} auth={auth} peerxy={undefined} dataChannel={dataChannel} />
-      )}
-      {(auth ? myWait === 9 : peerWait === 9) && segmentState === "ice" && (
-        <Ice key={`ice_${i}`} i={i} auth={auth} segmentState={segmentState} videoId={videoId} peerxy={peerxy} dataChannel={dataChannel} isRightCard={isRightCard} />
       )}
       {(auth ? myWait === 9 : peerWait === 9) && segmentState === "magnet" && (
         <Magnet key={`magnet_${i}`} i={i} auth={auth} peerxy={undefined} dataChannel={dataChannel} videoId={videoId} segmentState={segmentState} />
